Improve features

+++ app/src/main/java/com/example/minicapstone390/Controllers/SharedPreferenceHelper.java
@@ -1,7 +1,6 @@
 package com.example.minicapstone390.Controllers;
 
 import android.content.Context;
-import android.content.Intent;
 import android.content.SharedPreferences;
 
 public class SharedPreferenceHelper {
@@ -19,8 +18,10 @@ public class SharedPreferenceHelper {
         editor.apply();
     }
 
+    // Get theme mode
     public Boolean getTheme() { return sharedPreferences.getBoolean("themeMode", false); }
 
+    // Set score for a device
     public void setScore(float score, String id) {
         SharedPreferences.Editor editor = sharedPreferences.edit();
         assert editor != null;
@@ -28,5 +29,6 @@ public class SharedPreferenceHelper {
         editor.apply();
     }
 
-    public double getScore(String id) { return sharedPreferences.getFloat(id, 1); }
+    // Get score for a device
+    public float getScore(String id) { return sharedPreferences.getFloat(id, 1); }
 }
